requirements.txt
xlrd 
openpyxl
google-genai
